+++ HISTORY.rst
@@ -3,4 +3,4 @@ History
 
 v0.4.0dev
 +++++++++
-- Changes here...
+- OAuth1Session methods only return unicode strings. #55.